Revert "Add consign error to market white labels"
This reverts commit cb972eb448e3d8eb2643a3a3b12f219644c7479f.
Unfortunately, the edition's `acl_edit` can still be true when
the piece's `acl_edit` is false, so there is no good way on
the frontend right now to tell when to show the error message.

For now we'll assume submissions to marketplaces are all new
and ignore this edge case.